gnp-backend/requirements.txt
# Web Framework and Server
fastapi
uvicorn[standard]

# Database and ORM
sqlalchemy
asyncpg

# Google Gmail API Integration
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-auth

# Data Validation and Utilities
pydantic
python-multipart